backend_drf_inquiring/requirements.txt
asgiref==3.6.0
boto3==1.26.8
botocore==1.29.134
crispy-bootstrap5==0.6
Django==4.0.2
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-dotenv==1.4.2
django-environ==0.8.1
django-mysql==3.8.1
django-storages==1.13.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
jmespath==1.0.1
# Pillow==9.0.0
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
whitenoise==5.3.0
psycopg2-binary
